+++ packages/client/src/coordinator/connection/connection.ts
@@ -1,13 +1,7 @@
-import WebSocket from 'isomorphic-ws';
 import { StreamClient } from './client';
 import {
-  buildWsFatalInsight,
-  buildWsSuccessAfterFailureInsight,
-  postInsights,
-} from './insights';
-import {
   addConnectionEventListeners,
-  convertErrorToJson,
+  isCloseEvent,
   KnownCodes,
   randomId,
   removeConnectionEventListeners,
@@ -15,19 +9,12 @@ import {
   sleep,
 } from './utils';
 import type { LogLevel, StreamVideoEvent, UR } from './types';
-import type { ConnectedEvent, WSAuthMessage } from '../../gen/coordinator';
+import type {
+  ConnectedEvent,
+  ConnectionErrorEvent,
+  WSAuthMessage,
+} from '../../gen/coordinator';
 import { makeSafePromise, type SafePromise } from '../../helpers/promise';
-
-// Type guards to check WebSocket error type
-const isCloseEvent = (
-  res: WebSocket.CloseEvent | WebSocket.Data | WebSocket.ErrorEvent,
-): res is WebSocket.CloseEvent =>
-  (res as WebSocket.CloseEvent).code !== undefined;
-
-const isErrorEvent = (
-  res: WebSocket.CloseEvent | WebSocket.Data | WebSocket.ErrorEvent,
-): res is WebSocket.ErrorEvent =>
-  (res as WebSocket.ErrorEvent).error !== undefined;
 
 /**
  * StableWSConnection - A WS connection that reconnects upon failure.
@@ -50,7 +37,6 @@ export class StableWSConnection {
   // local vars
   connectionID?: string;
   private connectionOpenSafe?: SafePromise<ConnectedEvent>;
-  authenticationSent: boolean;
   consecutiveFailures: number;
   pingInterval: number;
   healthCheckTimeoutRef?: NodeJS.Timeout;
@@ -84,8 +70,6 @@ export class StableWSConnection {
     this.totalFailures = 0;
     /** We only make 1 attempt to reconnect at the same time.. */
     this.isConnecting = false;
-    /** True after the auth payload is sent to the server */
-    this.authenticationSent = false;
     /** To avoid reconnect if client is disconnected */
     this.isDisconnected = false;
     /** Boolean that indicates if the connection promise is resolved */
@@ -219,12 +203,9 @@ export class StableWSConnection {
    */
   _buildUrl = () => {
     const params = new URLSearchParams();
-    // const qs = encodeURIComponent(this.client._buildWSPayload(this.requestID));
-    // params.set('json', qs);
     params.set('api_key', this.client.key);
     params.set('stream-auth-type', this.client.getAuthType());
     params.set('X-Stream-Client', this.client.getUserAgent());
-    // params.append('authorization', this.client._getToken()!);
 
     return `${this.client.wsBaseURL}/connect?${params.toString()}`;
   };
@@ -243,21 +224,12 @@ export class StableWSConnection {
     this.isDisconnected = true;
 
     // start by removing all the listeners
-    if (this.healthCheckTimeoutRef) {
-      clearInterval(this.healthCheckTimeoutRef);
-    }
-    if (this.connectionCheckTimeoutRef) {
-      clearInterval(this.connectionCheckTimeoutRef);
-    }
+    clearInterval(this.healthCheckTimeoutRef);
+    clearInterval(this.connectionCheckTimeoutRef);
 
     removeConnectionEventListeners(this.onlineStatusChanged);
 
     this.isHealthy = false;
-
-    // remove ws handlers...
-    if (this.ws && this.ws.removeAllListeners) {
-      this.ws.removeAllListeners();
-    }
 
     let isClosedPromise: Promise<void>;
     // and finally close...
@@ -266,7 +238,7 @@ export class StableWSConnection {
     const { ws } = this;
     if (ws && ws.close && ws.readyState === ws.OPEN) {
       isClosedPromise = new Promise((resolve) => {
-        const onclose = (event: WebSocket.CloseEvent) => {
+        const onclose = (event: CloseEvent) => {
           this._log(
             `disconnect() - resolving isClosedPromise ${
               event ? 'with' : 'without'
@@ -308,14 +280,9 @@ export class StableWSConnection {
    * @return {ConnectAPIResponse<ConnectedEvent>} Promise that completes once the first health check message is received
    */
   async _connect() {
-    if (
-      this.isConnecting ||
-      (this.isDisconnected && this.client.options.enableWSFallback)
-    )
-      return; // simply ignore _connect if it's currently trying to connect
+    if (this.isConnecting) return; // simply ignore _connect if it's currently trying to connect
     this.isConnecting = true;
     this.requestID = randomId();
-    this.client.insightMetrics.connectionStartTimestamp = new Date().getTime();
     let isTokenReady = false;
     try {
       this._log(`_connect() - waiting for token`);
@@ -342,7 +309,8 @@ export class StableWSConnection {
         wsURL,
         requestID: this.requestID,
       });
-      this.ws = new WebSocket(wsURL);
+      const WS = this.client.options.WebSocketImpl ?? WebSocket;
+      this.ws = new WS(wsURL);
       this.ws.onopen = this.onopen.bind(this, this.wsID);
       this.ws.onclose = this.onclose.bind(this, this.wsID);
       this.ws.onerror = this.onerror.bind(this, this.wsID);
@@ -353,18 +321,6 @@ export class StableWSConnection {
       if (response) {
         this.connectionID = response.connection_id;
         this.client.resolveConnectionId?.(this.connectionID);
-        if (
-          this.client.insightMetrics.wsConsecutiveFailures > 0 &&
-          this.client.options.enableInsights
-        ) {
-          postInsights(
-            'ws_success_after_failure',
-            buildWsSuccessAfterFailureInsight(
-              this as unknown as StableWSConnection,
-            ),
-          );
-          this.client.insightMetrics.wsConsecutiveFailures = 0;
-        }
         return response;
       }
     } catch (err) {
@@ -372,16 +328,6 @@ export class StableWSConnection {
       this.isConnecting = false;
       // @ts-ignore
       this._log(`_connect() - Error - `, err);
-      if (this.client.options.enableInsights) {
-        this.client.insightMetrics.wsConsecutiveFailures++;
-        this.client.insightMetrics.wsTotalFailures++;
-
-        const insights = buildWsFatalInsight(
-          this as unknown as StableWSConnection,
-          convertErrorToJson(err as Error),
-        );
-        postInsights?.('ws_fatal', insights);
-      }
       this.client.rejectConnectionId?.(err);
       throw err;
     }
@@ -422,7 +368,7 @@ export class StableWSConnection {
       return;
     }
 
-    if (this.isDisconnected && this.client.options.enableWSFallback) {
+    if (this.isDisconnected) {
       this._log('_reconnect() - Abort (3) since disconnect() is called');
       return;
     }
@@ -518,12 +464,11 @@ export class StableWSConnection {
       },
     };
 
-    this.authenticationSent = true;
     this.ws?.send(JSON.stringify(authMessage));
     this._log('onopen() - onopen callback', { wsID });
   };
 
-  onmessage = (wsID: number, event: WebSocket.MessageEvent) => {
+  onmessage = (wsID: number, event: MessageEvent) => {
     if (this.wsID !== wsID) return;
 
     this._log('onmessage() - onmessage callback', { event, wsID });
@@ -538,7 +483,6 @@ export class StableWSConnection {
     if (!this.isResolved && data && data.type === 'connection.error') {
       this.isResolved = true;
       if (data.error) {
-        // @ts-expect-error - the types of _errorFromWSEvent are incorrect
         this.rejectPromise?.(this._errorFromWSEvent(data, false));
         return;
       }
@@ -584,7 +528,7 @@ export class StableWSConnection {
     this.scheduleConnectionCheck();
   };
 
-  onclose = (wsID: number, event: WebSocket.CloseEvent) => {
+  onclose = (wsID: number, event: CloseEvent) => {
     if (this.wsID !== wsID) return;
 
     this._log('onclose() - onclose callback - ' + event.code, { event, wsID });
@@ -594,11 +538,15 @@ export class StableWSConnection {
       // usually caused by invalid auth details
       const error = new Error(
         `WS connection reject with error ${event.reason}`,
-      ) as Error & WebSocket.CloseEvent;
+      );
 
+      // @ts-expect-error
       error.reason = event.reason;
+      // @ts-expect-error
       error.code = event.code;
+      // @ts-expect-error
       error.wasClean = event.wasClean;
+      // @ts-expect-error
       error.target = event.target;
 
       this.rejectPromise?.(error);
@@ -622,14 +570,14 @@ export class StableWSConnection {
     }
   };
 
-  onerror = (wsID: number, event: WebSocket.ErrorEvent) => {
+  onerror = (wsID: number, event: Event) => {
     if (this.wsID !== wsID) return;
 
     this.consecutiveFailures += 1;
     this.totalFailures += 1;
     this._setHealth(false);
     this.isConnecting = false;
-    this.rejectPromise?.(this._errorFromWSEvent(event));
+    this.rejectPromise?.(new Error(`WebSocket error: ${event}`));
     this._log(`onerror() - WS connection resulted into error`, { event });
 
     this._reconnect();
@@ -641,7 +589,6 @@ export class StableWSConnection {
    *
    * @param {boolean} healthy boolean indicating if the connection is healthy or not
    * @param {boolean} dispatchImmediately boolean indicating to dispatch event immediately even if the connection is unhealthy
-   *
    */
   _setHealth = (healthy: boolean, dispatchImmediately = false) => {
     if (healthy === this.isHealthy) return;
@@ -668,40 +615,31 @@ export class StableWSConnection {
 
   /**
    * _errorFromWSEvent - Creates an error object for the WS event
-   *
    */
-  _errorFromWSEvent = (
-    event: WebSocket.CloseEvent | WebSocket.Data | WebSocket.ErrorEvent,
+  private _errorFromWSEvent = (
+    event: CloseEvent | ConnectionErrorEvent,
     isWSFailure = true,
   ) => {
-    let code;
-    let statusCode;
-    let message;
+    let code: number;
+    let statusCode: number;
+    let message: string;
     if (isCloseEvent(event)) {
       code = event.code;
-      statusCode = 'unknown';
       message = event.reason;
+      statusCode = 0;
+    } else {
+      const { error } = event;
+      code = error.code;
+      message = error.message;
+      statusCode = error.StatusCode;
     }
 
-    if (isErrorEvent(event)) {
-      code = event.error.code;
-      statusCode = event.error.StatusCode;
-      message = event.error.message;
-    }
-
-    // Keeping this `warn` level log, to avoid cluttering of error logs from ws failures.
-    this._log(
-      `_errorFromWSEvent() - WS failed with code ${code}`,
-      { event },
-      'warn',
-    );
-
-    const error = new Error(
-      `WS failed with code ${code} and reason - ${message}`,
-    ) as Error & {
-      code?: string | number;
+    const msg = `WS failed with code: ${code} and reason: ${message}`;
+    this._log(msg, { event }, 'warn');
+    const error = new Error(msg) as Error & {
+      code?: number;
       isWSFailure?: boolean;
-      StatusCode?: string | number;
+      StatusCode?: number;
     };
     error.code = code;
     /**
@@ -723,7 +661,6 @@ export class StableWSConnection {
     this.wsID += 1;
 
     try {
-      this?.ws?.removeAllListeners();
       this?.ws?.close();
     } catch (e) {
       // we don't care
@@ -752,11 +689,8 @@ export class StableWSConnection {
    * Schedules a next health check ping for websocket.
    */
   scheduleNextPing = () => {
-    if (this.healthCheckTimeoutRef) {
-      clearTimeout(this.healthCheckTimeoutRef);
-    }
-
     // 30 seconds is the recommended interval (messenger uses this)
+    clearTimeout(this.healthCheckTimeoutRef);
     this.healthCheckTimeoutRef = setTimeout(() => {
       // send the healthcheck..., server replies with a health check event
       const data = [{ type: 'health.check', client_id: this.client.clientID }];
@@ -775,10 +709,7 @@ export class StableWSConnection {
    * to be reconnected.
    */
   scheduleConnectionCheck = () => {
-    if (this.connectionCheckTimeoutRef) {
-      clearTimeout(this.connectionCheckTimeoutRef);
-    }
-
+    clearTimeout(this.connectionCheckTimeoutRef);
     this.connectionCheckTimeoutRef = setTimeout(() => {
       const now = new Date();
       if (